ai-ml-service/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image processing
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0

# OCR
pytesseract>=0.3.10

# Machine Learning (compatible with Python 3.13)
scikit-learn>=1.4.0
scikit-image>=0.22.0

# Utilities
requests>=2.31.0
aiofiles>=23.2.1
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Async support
httpx>=0.25.0

# For future deep learning models (install separately if needed)
# tensorflow>=2.15.0
# torch>=2.1.0
# torchvision>=0.16.0
